# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIPD API Documentation",
        "description": "Dokumentasi API untuk aplikasi SIPD",
        "contact": {
            "email": "admin@sipd_new.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://sipd_new.test/api/",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/documents/applicant": {
            "get": {
                "tags": [
                    "Applicant"
                ],
                "operationId": "1b9ef096190207fb152d8120839ad209",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "properties": {
                                                "status": {
                                                    "type": "integer",
                                                    "example": 200
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "success get all project"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "code_project": {
                                                        "type": "string",
                                                        "example": "PRJ-20260827-WL0R"
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Kampung Nelayan Merah Putih"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Pengajuan izin analisis dampak pembangunan program Kampung Nelayan Merah Putih."
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 4
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Response Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Unprocessable Entity"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/documents/verificator": {
            "get": {
                "tags": [
                    "Verificator"
                ],
                "operationId": "9b9749026de3b6466a4ee448cb153a27",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "properties": {
                                                "status": {
                                                    "type": "integer",
                                                    "example": 200
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "success get all documents"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "code_project": {
                                                        "type": "string",
                                                        "example": "PRJ-20260827-WL0R"
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Kampung Nelayan Merah Putih"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Pengajuan izin analisis dampak pembangunan program Kampung Nelayan Merah Putih."
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 4
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Response Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Unprocessable Entity"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/documents": {
            "post": {
                "tags": [
                    "Applicant"
                ],
                "operationId": "1c3ad2906b68a58de1f603260ab271c4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/DocumentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Response Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "properties": {
                                                "status": {
                                                    "type": "integer",
                                                    "example": 200
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "success get all project"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "code_project": {
                                                        "type": "string",
                                                        "example": "PRJ-20260827-WL0R"
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Kampung Nelayan Merah Putih"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Pengajuan izin analisis dampak pembangunan program Kampung Nelayan Merah Putih."
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 4
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Response Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Unprocessable Entity"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DocumentRequest": {
                "title": "Document Request",
                "description": "",
                "required": [
                    "project_id",
                    "title",
                    "files[]"
                ],
                "properties": {
                    "project_id": {
                        "description": "ID Project terkait",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Judul Dokumen",
                        "type": "string",
                        "example": "Dokumen AMDAL Tahap 1"
                    },
                    "description": {
                        "description": "Deskripsi tambahan (opsional)",
                        "type": "string",
                        "example": "Pengajuan dokumen AMDAL untuk kawasan pesisir.",
                        "nullable": true
                    },
                    "document_type": {
                        "description": "Tipe Dokumen (opsional)",
                        "type": "string",
                        "example": "AMDAL",
                        "nullable": true
                    },
                    "files[]": {
                        "description": "Upload 1 atau lebih file (pdf, jpg, jpeg, png, docx. Max: 5MB per file)",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "binary"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Sanctum Token"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}